requirements.txt
pytest
numpy
pytesseract
pillow
